# optimize `mutable_copy` on `BigFloat` and `BigInt`