--- a/src/bigfloat.jl
+++ b/src/bigfloat.jl
@@ -1,5 +1,9 @@
 mutability(::Type{BigFloat}) = IsMutable()
-mutable_copy(x::BigFloat) = deepcopy(x)
+function mutable_copy(x::BigFloat)
+    d = x._d
+    d′ = GC.@preserve d unsafe_string(pointer(d), sizeof(d)) # creates a definitely-new String
+    return Base.MPFR._BigFloat(x.prec, x.sign, x.exp, d′)
+end
 
 @static if VERSION >= v"1.1.0-DEV.683"
     const _MPFRRoundingMode = Base.MPFR.MPFRRoundingMode