--- a/src/bigint.jl
+++ b/src/bigint.jl
@@ -1,4 +1,6 @@
 mutability(::Type{BigInt}) = IsMutable()
+# copied from `deepcopy_internal` implementation in Julia:
+# https://github.com/JuliaLang/julia/blob/7d41d1eb610cad490cbaece8887f9bbd2a775021/base/gmp.jl#L772
 mutable_copy(x::BigInt) = Base.GMP.MPZ.set(x)
 
 # zero