The `deepcopy` in Base creates an `IDDict` to keep track of already copied values:

https://github.com/JuliaLang/julia/blob/60e3e5527f0395d32fe2d59d3eb6c0a5cf59d354/base/deepcopy.jl#L26

However, in our case, this is not needed and we might as well just inline the definition of `deepcopy_internal` without the dictionary. This makes us rely on some of the internals of `MPFR` and `GMP` but this package already does that heavily so might as well go "all in" in that case.

On master the performance is:

```jl
julia> using MutableArithmetics, BenchmarkTools

julia> bf, bi = big"1.0", big"1";

julia> @btime MutableArithmetics.mutable_copy($bf);
  89.067 ns (5 allocations: 464 bytes)

julia> @btime MutableArithmetics.mutable_copy($bi);
  119.106 ns (4 allocations: 376 bytes)
```

while on this PR it is:

```jl
julia> bf, bi = big"1.0", big"1";

julia> @btime MutableArithmetics.mutable_copy($bf);
  18.507 ns (2 allocations: 112 bytes)

julia> @btime MutableArithmetics.mutable_copy($bi);
  63.083 ns (2 allocations: 40 bytes)
```