@KristofferC 2021-11-05
The fact that this package exists for performance and talks about bytes allocated etc in the README.