@blegat 2021-11-05
Thanks. Indeed, using `deepcopy` isn't optimal there. I noticed `copy` just returned the same instance but didn't investigate further. It's going to improve the performance of `@rewrite` and of packages using this function such as the packages in JuliaAlgebra.
Could you leave a comment in the implementation of BigFloat saying that this is copy-pasted from `deepcopy_internal` so that we know where to look if it needs to be updated at some point ?